requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.5
python-dotenv>=1.0.0
dj-database-url>=2.1.0
whitenoise>=6.5.0
requests>=2.31.0
PyJWT>=2.8.0
google-generativeai
pillow
